Extract parser and reflect to packages
